reorganize routes, consolidate EC2 controllers, delete console.logs and add comments
