+++ server/controllers/cookieController.js
@@ -1,22 +1,15 @@
 const cookieController = {};
-const User = require('../models/userModel');
 
+//set the SSID cookie based on user's _id in the database
 cookieController.setSSIDCookie = async (req, res, next) => {
   try {
-    // assigning the newUserID data to a var - DOES THE MANGO GENERATED ID COME WITH THIS?
     const userId = res.locals.newUser._id;
-    // IF I CANT GET THE ID LIKE THIS I WILL USE THIS METHOD (maybe need exec):
-    // const user = await User.find({ email: req.body.email });
-    // const userId = newUser[0]._id;
-    // console.log('inside cookie controller:, ', userId);
-    //creating cookie
     res.cookie('ssid', userId, {
       httpOnly: true,
     });
     res.locals.ssidCookie = userId;
     return next();
   } catch (err) {
-    // send to global error handler
     next({
       log: `Error in cookieController.setSSIDCookie. Details: ${err}`,
       message: { err: 'An error occurred in cookieController.setSSIDCookie' },
