+++ server/controllers/ec2/cpuMetricsController.js
@@ -5,22 +5,26 @@ const {
 
 const cpuMetricsController = {};
 
+// Get CPU metrics for all instances in a region
 cpuMetricsController.getCPUMetrics = async (req, res, next) => {
-  const { region } = req.query;
   const { instances } = res.locals.ec2Instances;
 
   const credentials = {
-    region: region,
+    region: req.query.region,
     credentials: res.locals.credentials,
   };
 
+  // Initiate StartTime to be 7 days before EndTime
+  const EndTime = new Date();
+  const StartTime = new Date(EndTime.getTime() - 7 * 24 * 60 * 60 * 1000);
+
+  // Create new instance of CloudWatchClient with user's region and credentials
   const cloudwatch = new CloudWatchClient(credentials);
 
   try {
-    const EndTime = new Date();
-    const StartTime = new Date(EndTime.getTime() - 7 * 24 * 60 * 60 * 1000);
     const queries = [];
 
+    // Generate AWS EC2 CPU metrics queries for each instance
     instances.forEach((instanceId, index) => {
       const id = index * instances.length + index + 1;
       queries.push(
@@ -108,37 +112,25 @@ cpuMetricsController.getCPUMetrics = async (req, res, next) => {
       MetricDataQueries: queries,
     };
 
+    // Send GetMetricDataCommand to AWS CloudWatch
     const command = new GetMetricDataCommand(input);
     const responses = await cloudwatch.send(command);
-    const chartData = responses.MetricDataResults.reduce((acc, curr) => {
+
+    //format data to be sent to frontend
+    const timestamps = responses.MetricDataResults[0].Timestamps;
+    const data = responses.MetricDataResults.reduce((acc, curr) => {
       if (!acc[curr.Label]) {
         acc[curr.Label] = {
           values: [],
-          timestamps: [],
+          timestamps: timestamps,
           instanceIds: instances,
         };
       }
-      acc[curr.Label].values = curr.Values;
-      acc[curr.Label].timestamps = curr.Timestamps;
+      acc[curr.Label].values.push(curr.Values);
       return acc;
     }, {});
 
-    // console.log('chartData', chartData);
-    res.locals.chartData = chartData;
-    console.log('res.locals.chartData', res.locals.chartData);
-    // res.locals.chartData.CPUCreditUsage = chartData;
-    // res.locals.chartData.CPUCreditBalance = chartData;
-    // res.locals.chartData.CPUSurplusCreditBalance = chartData;
-
-    /*
-    chartData = {
-        CPUUtilization: {values: [], timestamps: [], instanceIds: []},
-        CPUCreditUsage: {values: [], timestamps: [], instanceIds: []},
-        CPUCreditBalance: {values: [], timestamps: [], instanceIds: []},
-        CPUSurplusCreditBalance: {values: [], timestamps: [], instanceIds: []},
-    };
-    */
-
+    res.locals.chartData = data;
     return next();
   } catch (error) {
     console.error(error);
