+++ server/controllers/ec2/instancesController.js
@@ -1,5 +1,6 @@
 const { EC2Client, DescribeInstancesCommand } = require('@aws-sdk/client-ec2');
 
+//retrieve ids from all ec2 instances in a region
 module.exports = {
   getInstances: async (req, res, next) => {
     const info = {
@@ -7,9 +8,13 @@ module.exports = {
       credentials: res.locals.credentials,
     };
 
+    //create new instance of EC2Client with user's region and credentials
     const ec2Client = new EC2Client(info);
+
     try {
       const data = await ec2Client.send(new DescribeInstancesCommand({}));
+
+      //extract instance ids from data
       const instances = data.Reservations;
       const instanceIds = instances.map((instance) => {
         return instance.Instances[0].InstanceId;
